requirements.txt
pandas
numpy
sentence-transformers
faiss-cpu
streamlit
openai
langchain
